requirements.txt
requests
pytest
pytest-bdd
pytest-html
python-dotenv
pyotp
